+++ src/controllers/post.controller.ts
@@ -4,6 +4,7 @@ import Controller from '../interfaces/controller.interface';
 import Post from '../interfaces/post.interface';
 import postModel from '../models/post.model';
 import PostNotFoundException from '../exceptions/PostNotFoundException';
+import NotAuthorizedException from '../exceptions/NotAuthorizedException';
 import postValidator from '../validators/post.validator';
 import validationMiddleware from '../middleware/validation.middleware';
 import authMiddleware from '../middleware/auth.middleware';
@@ -20,7 +21,6 @@ class PostController implements Controller {
 	}
 
 	private initializeRoutes() {
-		this.router.get(this.path, this.getAllPosts);
 		this.router.get(`${this.path}/:id`, this.getPostById);
 
 		this.router
@@ -81,25 +81,15 @@ class PostController implements Controller {
 				throw new PostNotFoundException(id);
 			}
 
+			if (post.author.toString() !== req.user._id.toString()) {
+				throw new NotAuthorizedException();
+			}
+
 			deleteFile(post.image);
 			await this.post.findByIdAndDelete(id);
 
 			const message = 'Post deleted successfully';
 			res.status(200).json({ message });
-		} catch (err) {
-			next(err);
-		}
-	};
-
-	private getAllPosts = async (
-		req: Request,
-		res: Response,
-		next: NextFunction
-	) => {
-		try {
-			const posts = await this.post.find();
-			const message = 'Posts fetched successfully';
-			res.status(200).json({ message, posts });
 		} catch (err) {
 			next(err);
 		}
@@ -141,6 +131,10 @@ class PostController implements Controller {
 
 			if (!post) {
 				throw new PostNotFoundException(id);
+			}
+
+			if (post.author.toString() !== req.user._id.toString()) {
+				throw new NotAuthorizedException();
 			}
 
 			const message = 'Post updated successfully';
